+++ .gitignore
@@ -12,5 +12,5 @@ config.json
 *.xz
 video/dav1d-1.4.2
 video/ffmpeg-7.0.1
-
+build/
 serial/*.bin
