+++ serial/build.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+mkdir -p serial
+cd ../serial
+g++ demux.cpp SerialPort.cpp -lgdi32 -o serial.exe
+mv serial.exe ../build/serial
+cd ../build
