+++ video/FileStreamSource.js
@@ -1,11 +1,18 @@
 const VideoSource = require("./VideoSource");
+const { log } = require("../debug");
 const { spawn } = require("child_process");
 const { Readable } = require("stream");
 const fs = require("fs");
 const path = require("path");
 
-
-const ffmpegPath = path.join(__dirname, "ffmpeg-7.0.1", "ffmpeg.exe");
+const ffmpegPath = path.join(
+  __dirname,
+  "..",
+  "build",
+  "video",
+  "ffmpeg-7.0.1",
+  "ffmpeg.exe"
+);
 
 /**
  * A class to play a local file as a video source
@@ -23,10 +30,7 @@ class FileStreamSource extends VideoSource {
    */
   constructor(file, isVideo1, options) {
     //call the VideoSource constructor with the name as the file name
-    super(
-      isVideo1 ? "LiveVideo1" : "LiveVideo2",
-      fs.createReadStream(file)
-    );
+    super(isVideo1 ? "LiveVideo1" : "LiveVideo2", fs.createReadStream(file));
 
     console.log("File stream source created" + isVideo1 + " " + file);
